requirements.txt
faker
xlsxwriter
python-docx
python-pptx
aspose-diagram
pyyaml
reportlab
tabulate
pyzipper
requests
pysftp
rpyc
fabric
decorator
dnslib
dnspython
ping3
dnslib
flask
graphql-server
jpype1
dict2xml
pycrypto
pycryptodome
websockets
grpcio
grpcio-tools
boto3
wonderwords
graphene
ariadne